requirements.txt
alembic==1.8.1
async-timeout==4.0.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
DBUtils==3.0.2
Deprecated==1.2.13
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
pycparser==2.21
pycryptodome==3.15.0
PyMySQL==1.0.2
pyparsing==3.0.9
redis==4.3.4
SQLAlchemy==1.4.35
Werkzeug==2.2.2
wrapt==1.14.1

uWSGI==2.0.20